requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pybit==5.6.2
requests==2.31.0
redis==5.0.1
hiredis==2.3.2

# Web Framework
fastapi==0.110.0
uvicorn[standard]==0.27.0
jinja2==3.1.3
python-multipart==0.0.6
httpx==0.26.0
pydantic==2.5.3

# Async Support
aioredis==2.0.1
asyncio==3.4.3
aiofiles==23.2.1

# Flask (legacy, optional)
flask==3.0.0
flask-cors==4.0.0

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Monitoring & Logging
prometheus-client==0.19.0
structlog==24.1.0

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Utilities
schedule==1.2.0
python-json-logger==2.0.7

# Telegram Notifications
telethon==1.34.0